pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "wikipedia>=1.4.0",
]
